refactor: [#265] WorkoutManager와 MatricsIndicator를 분리
- 세션의 진행, 멈춤, 빌드와 관련된 코드 -> WorkoutManager
- 경기의 데이터에 관련한 코드 -> MatricsIndicator
